requirements.txt
Django==1.11.17
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
psycopg2==2.7.6.1
requests==2.20.1